Conflicts:
	.classpath
	.settings/org.eclipse.jdt.core.prefs
	.settings/org.eclipse.wst.common.component
	src/lib/JarManagerAPI.jar
	src/main/webapp/WEB-INF/views/jsp/home.jsp

+++ src/main/java/tw/com/jarmanager/q2d/web/mode/QueueConnectionFactory.java
@@ -4,7 +4,7 @@ public class QueueConnectionFactory {
 	private String host;
 	private String virtualHost;
 	private String password;
-	private long port;
+	private int port;
 	private String username;
 
 	public String getHost() {
@@ -31,11 +31,11 @@ public class QueueConnectionFactory {
 		this.password = password;
 	}
 
-	public long getPort() {
+	public int getPort() {
 		return port;
 	}
 
-	public void setPort(long port) {
+	public void setPort(int port) {
 		this.port = port;
 	}
